some chenges

+++ routes/web.php
@@ -32,9 +32,9 @@ Route::get('/rec.ipe', [WelcomeRecipesController::class, 'welcomeRecipes'])->nam
 Route::get('/creatrecipe', [WelcomeRecipesController::class, 'creatrecipe'])->name('recipes.creat');
 Route::post('/creatrecipe', [WelcomeRecipesController::class, 'store'])->name('recipes.store');
 Route::get('/user_recipe/{user_id}/{id}', [WelcomeRecipesController::class, 'show'])->name('user.show');
-Route::get('/user_recipe_edit/{userrecipe}', [WelcomeRecipesController::class, 'edit'])->middleware('auth')->name('user.edit');
-Route::put('/user_recipe_edit/{userrecipe}', [WelcomeRecipesController::class, 'update'])->name('user.update');
-Route::delete('/user_recipe_edit/{userrecipe}', [WelcomeRecipesController::class, 'destroy'])->name('user.destroy');
+Route::get('/user_recipe_edit/{user_id}/{userrecipe}', [WelcomeRecipesController::class, 'edit'])->name('user.edit');
+Route::put('/user_recipe_edit/{user_id}/{userrecipe}', [WelcomeRecipesController::class, 'update'])->name('user.update');
+Route::delete('/user_recipe_edit/{user_id}/{userrecipe}', [WelcomeRecipesController::class, 'destroy'])->name('user.destroy');
